- Created all neccessary    directories(Controllers,Repositories,DTO,Services)    - In DTO directory added entities dtos    - I break down the branck into simple one.
