--- backend/CertificateManagerAPIs/Controllers/SupplierController.cs
@@ -1,55 +1,0 @@
-﻿using CertificateManagerAPIs.DTO;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.EntityFrameworkCore;
-
-namespace CertificateManagerAPIs.Controllers
-{
-    [ApiController]
-    [Route("api/[controller]")]
-    public class SupplierController : Controller
-    {
-        private readonly CertificatedbContext _context;
-        public SupplierController(CertificatedbContext context)
-        {
-            _context = context;
-        }
-
-        [HttpGet]
-        public async Task<ActionResult<IEnumerable<SupplierDto>>> GetSuppliers()
-        {
-            var suppliers = await _context.Suppliers.ToListAsync();
-
-            var supplierDtos = suppliers.Select(s => new SupplierDto
-            {
-                SupplierId = s.SupplierId,
-                SupplierName = s.SupplierName,
-                SupplierIndex = s.SupplierIndex,
-                City = s.City
-            }).ToList();
-
-            return Ok(supplierDtos);
-        }
-
-        [HttpGet("{id}")]
-        public async Task<ActionResult<SupplierDto>> GetSupplier(int id)
-        {
-            var supplier = await _context.Suppliers.FindAsync(id);
-
-            if (supplier == null)
-            {
-                return NotFound();
-            }
-
-            var supplierDto = new SupplierDto
-            {
-                SupplierId = supplier.SupplierId,
-                SupplierName = supplier.SupplierName,
-                SupplierIndex = supplier.SupplierIndex,
-                City = supplier.City
-            };
-
-            return Ok(supplierDto);
-        }
-
-    }
-}
