+++ backend/CertificateManagerAPIs/DTO/CertificateDto.cs
@@ -1,14 +1,15 @@
 ﻿namespace CertificateManagerAPIs.DTO
 {
-
     public class CertificateDto
     {
-        public int Id { get; set; }
-        public string Type { get; set; }
+
+        public SupplierDto Supplier { get; set; } = null!;
+        public string Type { get; set; } = null!;
+
         public DateTime ValidFrom { get; set; }
+
         public DateTime ValidTo { get; set; }
-        public SupplierDto Supplier { get; set; }
+        public int Id { get; set; }
     }
 
 }
-
