+++ backend/CertificateManagerAPIs/DTO/CommentDto.cs
@@ -2,7 +2,6 @@
 {
     public class CommentDto
     {
-
         public int Id { get; set; }
 
         public int CertificateId { get; set; }
@@ -10,6 +9,5 @@
         public int UserId { get; set; }
 
         public string UserComment { get; set; } = null!;
-
     }
 }
