+++ backend/CertificateManagerAPIs/DTO/SupplierDto.cs
@@ -6,9 +6,8 @@
 
         public string SupplierName { get; set; } = null!;
 
-        public int? SupplierIndex { get; set; }
+        public int SupplierIndex { get; set; }
 
         public string City { get; set; } = null!;
     }
 }
-
