+++ backend/CertificateManagerAPIs/DTO/UserDto.cs
@@ -1,0 +1,17 @@
+﻿namespace CertificateManagerAPIs.DTO
+{
+    public class UserDto
+    {
+        public int UserId { get; set; }
+
+        public string Name { get; set; } = null!;
+        public string FirstName { get; set; } = null!;
+
+        public string Department { get; set; }
+
+        public string Plant { get; set; }
+        public string Email { get; set; } = null!;
+
+    }
+}
+
