--- backend/CertificateManagerAPIs/Repository/CertificateRepository.cs
@@ -1,6 +1,0 @@
-﻿namespace CertificateManagerAPIs.Repository
-{
-    public class CertificateRepository
-    {
-    }
-}
