--- backend/CertificateManagerAPIs/Repository/ICertificateRepository.cs
@@ -1,6 +1,0 @@
-﻿namespace CertificateManagerAPIs.Repository
-{
-    public interface ICertificateRepository
-    {
-    }
-}
